Brians refactoring
