+++ .gitignore
@@ -1,0 +1,5 @@
+/bin/
+/build/
+/classes/
+*.class
+*.zip
